@@ -5,8 +5,8 @@ void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
   const MyApp({
-    Key key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -24,15 +24,15 @@ class MyApp extends StatelessWidget {
 class HomePage extends StatelessWidget {
   /// Creates a [HomePage].
   const HomePage({
-    Key key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return const Scaffold(
       body: SafeArea(
         child: Column(
-          children: const <Widget>[
+          children: <Widget>[
             Expanded(child: _Body(axis: Axis.vertical)),
             Expanded(child: _Body(axis: Axis.horizontal)),
           ],
@@ -46,9 +46,8 @@ class HomePage extends StatelessWidget {
 class _Body extends StatelessWidget {
   /// Creates a [_Body].
   const _Body({
-    Key key,
-    this.axis,
-  }) : super(key: key);
+    required this.axis,
+  });
 
   final Axis axis;
 
@@ -72,9 +71,8 @@ class _Body extends StatelessWidget {
 
 class _Gap extends StatelessWidget {
   const _Gap({
-    Key key,
-    this.color,
-  }) : super(key: key);
+    required this.color,
+  });
 
   final Color color;
 
