@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
 class RenderGap extends RenderBox {
@@ -40,7 +39,7 @@ class RenderGap extends RenderBox {
   }
 
   Axis? get _direction {
-    final AbstractNode? parentNode = parent;
+    final parentNode = parent;
     if (parentNode is RenderFlex) {
       return parentNode.direction;
     } else {
